+++ test/clear-stale/remove-one-field.ts
@@ -5,6 +5,7 @@ export const tests: ReadonlyArray<OneTest> = [
     name: "remove one field",
     normMap: { myid: { id: "myid", name: "foo" } },
     staleBefore: { myid: new Set(["name", "age"]) },
-    staleAfter: { myid: new Set(["age"]) }
+    staleAfter: { myid: new Set(["age"]) },
+    clearedAfter: { myid: new Set(["name"]) }
   }
 ];